#1036: Refactor docker.js to be more testable and maintainable
