+++ lib/logger.js
@@ -38,7 +38,7 @@ const logLevels = {
  * @param {...Any} [values] Values to be passed `utils.format()`
  * @example
  * // Log an error message
- * lando.log.error('This is an err with details %j', err);
+ * lando.log.error('This is an err with details %s', err);
  */
 /**
  * Logs an info message.
