+++ lib/promise.js
@@ -32,7 +32,7 @@ const retry = (fn, {max = 5, backoff = 500} = {}) => Promise.resolve().then(() =
       if (counter <= max) {
         return Promise.delay(backoff * counter).then(() => rec(counter + 1));
       } else {
-        throw new Error(format('Failed after %s retries with backoff %s and error %j', max, backoff, err));
+        throw new Error(format('Failed after %s retries with backoff %s and error %s', max, backoff, err));
       }
     }));
